requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
APScheduler==3.10.4
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
pandas==2.2.2
python-dateutil==2.9.0.post0
jinja2==3.1.4
streamlit==1.39.0
google-api-python-client==2.139.0
serpapi==0.1.5
openai==1.51.2
